Primeiro endpoint feito

+++ src/main/java/com/linked/model/ContaCorrente.java
@@ -1,18 +1,21 @@
 package com.linked.model;
 
+import java.math.BigDecimal;
+
 public class ContaCorrente {
 	private int id;
 	private int id_correntista;
 	private int id_agencia;
-	private double limite;
-	private double saldo;
+	private BigDecimal limite;
+	private BigDecimal saldo;
 	private String ativa;
 
 	public ContaCorrente() {
 		super();
 	}
 
-	public ContaCorrente(int id, int id_correntista, int id_agencia, double limite, double saldo, String ativa) {
+	public ContaCorrente(int id, int id_correntista, int id_agencia, BigDecimal limite, BigDecimal saldo,
+			String ativa) {
 		super();
 		this.id = id;
 		this.id_correntista = id_correntista;
@@ -46,19 +49,19 @@ public class ContaCorrente {
 		this.id_agencia = id_agencia;
 	}
 
-	public double getLimite() {
+	public BigDecimal getLimite() {
 		return limite;
 	}
 
-	public void setLimite(double limite) {
+	public void setLimite(BigDecimal limite) {
 		this.limite = limite;
 	}
 
-	public double getSaldo() {
+	public BigDecimal getSaldo() {
 		return saldo;
 	}
 
-	public void setSaldo(double saldo) {
+	public void setSaldo(BigDecimal saldo) {
 		this.saldo = saldo;
 	}
